Starting a refactor to cut the huge files into chunks
