+++ contracts/mixnet/src/bonding_gateways/mod.rs
@@ -1,0 +1,1 @@
+pub mod transactions;
