+++ contracts/mixnet/src/bonding_mixnodes/mod.rs
@@ -1,0 +1,1 @@
+pub mod transactions;
