+++ contracts/mixnet/src/lib.rs
@@ -1,10 +1,15 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 
+pub mod bonding_gateways;
+pub mod bonding_mixnodes;
 pub mod contract;
+pub mod delegating_mixnodes;
 pub mod error;
-pub(crate) mod helpers;
+pub mod helpers;
+pub mod mixnet_params;
 pub mod queries;
+mod rewards;
 pub mod state;
 pub(crate) mod storage;
 pub mod support;
