+++ contracts/mixnet/src/mixnet_params/mod.rs
@@ -1,0 +1,1 @@
+pub mod transactions;
