+++ contracts/mixnet/src/rewards/mod.rs
@@ -1,0 +1,1 @@
+pub mod transactions;
