+++ contracts/mixnet/src/transactions.rs
@@ -1,18 +1,11 @@
 // Copyright 2021 - Nym Technologies SA <contact@nymtech.net>
 // SPDX-License-Identifier: Apache-2.0
 use crate::error::ContractError;
-use crate::helpers::{calculate_epoch_reward_rate, scale_reward_by_uptime, Delegations};
-use crate::queries;
 use crate::storage::*;
 use config::defaults::DENOM;
-use cosmwasm_std::{
-    attr, coins, BankMsg, Coin, Decimal, DepsMut, Env, MessageInfo, Response, StdResult, Uint128,
-};
-use cosmwasm_storage::ReadonlyBucket;
-use mixnet_contract::mixnode::NodeRewardParams;
-use mixnet_contract::{
-    Gateway, GatewayBond, IdentityKey, Layer, MixNode, MixNodeBond, RawDelegationData, StateParams,
-};
+use cosmwasm_std::{coins, BankMsg, Coin, DepsMut, Env, MessageInfo, Response};
+
+use mixnet_contract::{IdentityKey, Layer, RawDelegationData};
 
 pub(crate) const OLD_DELEGATIONS_CHUNK_SIZE: usize = 500;
 
@@ -22,743 +15,47 @@ pub(crate) const MINIMUM_BLOCK_AGE_FOR_REWARDING: u64 = 17280;
 // approximately 30min (assuming 5s per block)
 pub(crate) const MAX_REWARDING_DURATION_IN_BLOCKS: u64 = 360;
 
-fn total_delegations(delegations_bucket: ReadonlyBucket<RawDelegationData>) -> StdResult<Coin> {
-    Ok(Coin::new(
-        Delegations::new(delegations_bucket)
-            .fold(0, |acc, x| acc + x.delegation_data.amount.u128()),
-        DENOM,
-    ))
-}
-
-fn validate_mixnode_bond(bond: &[Coin], minimum_bond: Uint128) -> Result<(), ContractError> {
-    // check if anything was put as bond
-    if bond.is_empty() {
-        return Err(ContractError::NoBondFound);
-    }
-
-    if bond.len() > 1 {
-        return Err(ContractError::MultipleDenoms);
-    }
-
-    // check that the denomination is correct
-    if bond[0].denom != DENOM {
-        return Err(ContractError::WrongDenom {});
-    }
-
-    // check that we have at least MIXNODE_BOND coins in our bond
-    if bond[0].amount < minimum_bond {
-        return Err(ContractError::InsufficientMixNodeBond {
-            received: bond[0].amount.into(),
-            minimum: minimum_bond.into(),
-        });
-    }
-
-    Ok(())
-}
-
-pub(crate) fn try_add_mixnode(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    mix_node: MixNode,
-) -> Result<Response, ContractError> {
-    let sender_bytes = info.sender.as_bytes();
-
-    // if the client has an active bonded gateway, don't allow mixnode bonding
-    if gateways_owners_read(deps.storage)
-        .may_load(sender_bytes)?
-        .is_some()
-    {
-        return Err(ContractError::AlreadyOwnsGateway);
-    }
-
-    let mut was_present = false;
-    // if the client has an active mixnode with a different identity, don't allow bonding
-    if let Some(existing_node) = mixnodes_owners_read(deps.storage).may_load(sender_bytes)? {
-        if existing_node != mix_node.identity_key {
-            return Err(ContractError::AlreadyOwnsMixnode);
-        }
-        was_present = true
-    }
-
-    // check if somebody else has already bonded a mixnode with this identity
-    if let Some(existing_bond) =
-        mixnodes_read(deps.storage).may_load(mix_node.identity_key.as_bytes())?
-    {
-        if existing_bond.owner != info.sender {
-            return Err(ContractError::DuplicateMixnode {
-                owner: existing_bond.owner,
-            });
-        }
-    }
-
-    let minimum_bond = read_state_params(deps.storage).minimum_mixnode_bond;
-    validate_mixnode_bond(&info.funds, minimum_bond)?;
-
-    let layer_distribution = queries::query_layer_distribution(deps.as_ref());
-    let layer = layer_distribution.choose_with_fewest();
-
-    let mut bond = MixNodeBond::new(
-        info.funds[0].clone(),
-        info.sender.clone(),
-        layer,
-        env.block.height,
-        mix_node,
-        None,
-    );
-
-    // this might potentially require more gas if a significant number of delegations was there
-    let delegations_bucket = mix_delegations_read(deps.storage, &bond.mix_node.identity_key);
-    let existing_delegation = total_delegations(delegations_bucket)?;
-    bond.total_delegation = existing_delegation;
-
-    let identity = bond.identity();
-
-    mixnodes(deps.storage).save(identity.as_bytes(), &bond)?;
-    mixnodes_owners(deps.storage).save(sender_bytes, identity)?;
-    increment_layer_count(deps.storage, bond.layer)?;
-
-    let attributes = vec![attr("overwritten", was_present)];
-    Ok(Response {
-        submessages: Vec::new(),
-        messages: Vec::new(),
-        attributes,
-        data: None,
-    })
-}
-
-pub(crate) fn try_remove_mixnode(
-    deps: DepsMut,
-    info: MessageInfo,
-) -> Result<Response, ContractError> {
-    let sender_bytes = info.sender.as_bytes();
-
-    // try to find the identity of the sender's node
-    let mix_identity = match mixnodes_owners_read(deps.storage).may_load(sender_bytes)? {
-        Some(identity) => identity,
-        None => return Err(ContractError::NoAssociatedMixNodeBond { owner: info.sender }),
-    };
-
-    // get the bond, since we found associated identity, the node MUST exist
-    let mixnode_bond = mixnodes_read(deps.storage).load(mix_identity.as_bytes())?;
-
-    // send bonded funds back to the bond owner
-    let messages = vec![BankMsg::Send {
-        to_address: info.sender.as_str().to_owned(),
-        amount: vec![mixnode_bond.bond_amount()],
-    }
-    .into()];
-
-    // remove the bond from the list of bonded mixnodes
-    mixnodes(deps.storage).remove(mix_identity.as_bytes());
-    // remove the node ownership
-    mixnodes_owners(deps.storage).remove(sender_bytes);
-    // decrement layer count
-    decrement_layer_count(deps.storage, mixnode_bond.layer)?;
-
-    // log our actions
-    let attributes = vec![attr("action", "unbond"), attr("mixnode_bond", mixnode_bond)];
-
-    Ok(Response {
-        submessages: Vec::new(),
-        messages,
-        attributes,
-        data: None,
-    })
-}
-
-fn validate_gateway_bond(bond: &[Coin], minimum_bond: Uint128) -> Result<(), ContractError> {
-    // check if anything was put as bond
-    if bond.is_empty() {
-        return Err(ContractError::NoBondFound);
-    }
-
-    if bond.len() > 1 {
-        return Err(ContractError::MultipleDenoms);
-    }
-
-    // check that the denomination is correct
-    if bond[0].denom != DENOM {
-        return Err(ContractError::WrongDenom {});
-    }
-
-    // check that we have at least 100 coins in our bond
-    if bond[0].amount < minimum_bond {
-        return Err(ContractError::InsufficientGatewayBond {
-            received: bond[0].amount.into(),
-            minimum: minimum_bond.into(),
-        });
-    }
-
-    Ok(())
-}
-
-pub(crate) fn try_add_gateway(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    gateway: Gateway,
-) -> Result<Response, ContractError> {
-    let sender_bytes = info.sender.as_bytes();
-
-    // if the client has an active bonded mixnode, don't allow gateway bonding
-    if mixnodes_owners_read(deps.storage)
-        .may_load(sender_bytes)?
-        .is_some()
-    {
-        return Err(ContractError::AlreadyOwnsMixnode);
-    }
-
-    let mut was_present = false;
-    // if the client has an active gateway with a different identity, don't allow bonding
-    if let Some(existing_node) = gateways_owners_read(deps.storage).may_load(sender_bytes)? {
-        if existing_node != gateway.identity_key {
-            return Err(ContractError::AlreadyOwnsGateway);
-        }
-        was_present = true
-    }
-
-    // check if somebody else has already bonded a gateway with this identity
-    if let Some(existing_bond) =
-        gateways_read(deps.storage).may_load(gateway.identity_key.as_bytes())?
-    {
-        if existing_bond.owner != info.sender {
-            return Err(ContractError::DuplicateGateway {
-                owner: existing_bond.owner,
-            });
-        }
-    }
-
-    let minimum_bond = read_state_params(deps.storage).minimum_gateway_bond;
-    validate_gateway_bond(&info.funds, minimum_bond)?;
-
-    let bond = GatewayBond::new(
-        info.funds[0].clone(),
-        info.sender.clone(),
-        env.block.height,
-        gateway,
-    );
-
-    let identity = bond.identity();
-    gateways(deps.storage).save(identity.as_bytes(), &bond)?;
-    gateways_owners(deps.storage).save(sender_bytes, identity)?;
-    increment_layer_count(deps.storage, Layer::Gateway)?;
-
-    let attributes = vec![attr("overwritten", was_present)];
-    Ok(Response {
-        submessages: Vec::new(),
-        messages: Vec::new(),
-        attributes,
-        data: None,
-    })
-}
-
-pub(crate) fn try_remove_gateway(
-    deps: DepsMut,
-    info: MessageInfo,
-) -> Result<Response, ContractError> {
-    let sender_bytes = info.sender.as_str().as_bytes();
-
-    // try to find the identity of the sender's node
-    let gateway_identity = match gateways_owners_read(deps.storage).may_load(sender_bytes)? {
-        Some(identity) => identity,
-        None => return Err(ContractError::NoAssociatedGatewayBond { owner: info.sender }),
-    };
-
-    // get the bond, since we found associated identity, the node MUST exist
-    let gateway_bond = gateways_read(deps.storage).load(gateway_identity.as_bytes())?;
-
-    // send bonded funds back to the bond owner
-    let messages = vec![BankMsg::Send {
-        to_address: info.sender.as_str().to_owned(),
-        amount: vec![gateway_bond.bond_amount()],
-    }
-    .into()];
-
-    // remove the bond from the list of bonded gateways
-    gateways(deps.storage).remove(gateway_identity.as_bytes());
-    // remove the node ownership
-    gateways_owners(deps.storage).remove(sender_bytes);
-    // decrement layer count
-    decrement_layer_count(deps.storage, Layer::Gateway)?;
-
-    // log our actions
-    let attributes = vec![
-        attr("action", "unbond"),
-        attr("address", info.sender),
-        attr("gateway_bond", gateway_bond),
-    ];
-
-    Ok(Response {
-        submessages: Vec::new(),
-        messages,
-        attributes,
-        data: None,
-    })
-}
-
-pub(crate) fn try_update_state_params(
-    deps: DepsMut,
-    info: MessageInfo,
-    params: StateParams,
-) -> Result<Response, ContractError> {
-    // note: In any other case, I wouldn't have attempted to unwrap this result, but in here
-    // if we fail to load the stored state we would already be in the undefined behaviour land,
-    // so we better just blow up immediately.
-    let mut state = config_read(deps.storage).load()?;
-
-    // check if this is executed by the owner, if not reject the transaction
-    if info.sender != state.owner {
-        return Err(ContractError::Unauthorized);
-    }
-
-    if params.mixnode_bond_reward_rate < Decimal::one() {
-        return Err(ContractError::DecreasingMixnodeBondReward);
-    }
-
-    if params.mixnode_delegation_reward_rate < Decimal::one() {
-        return Err(ContractError::DecreasingMixnodeDelegationReward);
-    }
-
-    // note: rewarded_set = active_set + idle_set
-    // hence rewarded set must always be bigger than (or equal to) the active set
-    if params.mixnode_rewarded_set_size < params.mixnode_active_set_size {
-        return Err(ContractError::InvalidActiveSetSize);
-    }
-
-    // if we're updating epoch length, recalculate rewards for mixnodes
-    if state.params.epoch_length != params.epoch_length {
-        state.mixnode_epoch_bond_reward =
-            calculate_epoch_reward_rate(params.epoch_length, params.mixnode_bond_reward_rate);
-        state.mixnode_epoch_delegation_reward =
-            calculate_epoch_reward_rate(params.epoch_length, params.mixnode_delegation_reward_rate);
-    } else {
-        // if mixnode rewards changed, recalculate respective values
-        if state.params.mixnode_bond_reward_rate != params.mixnode_bond_reward_rate {
-            state.mixnode_epoch_bond_reward =
-                calculate_epoch_reward_rate(params.epoch_length, params.mixnode_bond_reward_rate);
-        }
-        if state.params.mixnode_delegation_reward_rate != params.mixnode_delegation_reward_rate {
-            state.mixnode_epoch_delegation_reward = calculate_epoch_reward_rate(
-                params.epoch_length,
-                params.mixnode_delegation_reward_rate,
-            );
-        }
-    }
-
-    state.params = params;
-
-    config(deps.storage).save(&state)?;
-
-    Ok(Response::default())
-}
-
-// Note: this function is designed to work with only a single validator entity distributing rewards
-// The main purpose of this function is to update `latest_rewarding_interval_nonce` which
-// will trigger a different seed selection for the pseudorandom generation of the "demanded" set of mixnodes.
-pub(crate) fn try_begin_mixnode_rewarding(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    rewarding_interval_nonce: u32,
-) -> Result<Response, ContractError> {
-    let mut state = config_read(deps.storage).load()?;
-
-    // check if this is executed by the permitted validator, if not reject the transaction
-    if info.sender != state.rewarding_validator_address {
-        return Err(ContractError::Unauthorized);
-    }
-
-    // check whether sufficient number of blocks already elapsed since the previous rewarding happened
-    // (this implies the validator responsible for rewarding in the previous interval did not call
-    // `try_finish_mixnode_rewarding` - perhaps they crashed or something. Regardless of the reason
-    // it shouldn't prevent anyone from distributing rewards in the following interval)
-    // Do note, however, that calling `try_finish_mixnode_rewarding` is crucial as otherwise the
-    // "demanded" set won't get updated on the validator API side
-    if state.rewarding_in_progress
-        && state.rewarding_interval_starting_block + MAX_REWARDING_DURATION_IN_BLOCKS
-            > env.block.height
-    {
-        return Err(ContractError::RewardingInProgress);
-    }
-
-    // make sure the validator is in sync with the contract state
-    if rewarding_interval_nonce != state.latest_rewarding_interval_nonce + 1 {
-        return Err(ContractError::InvalidRewardingIntervalNonce {
-            received: rewarding_interval_nonce,
-            expected: state.latest_rewarding_interval_nonce + 1,
-        });
-    }
-
-    state.rewarding_interval_starting_block = env.block.height;
-    state.latest_rewarding_interval_nonce = rewarding_interval_nonce;
-    state.rewarding_in_progress = true;
-
-    config(deps.storage).save(&state)?;
-
-    let mut response = Response::new();
-    response.add_attribute(
-        "rewarding interval nonce",
-        rewarding_interval_nonce.to_string(),
-    );
-    Ok(response)
-}
-
-// Note: if any changes are made to this function or anything it is calling down the stack,
-// for example delegation reward distribution, the gas limits must be retested and both
-// validator-api/src/rewarding/mod.rs::{MIXNODE_REWARD_OP_BASE_GAS_LIMIT, PER_MIXNODE_DELEGATION_GAS_INCREASE}
-// must be updated appropriately.
-pub(crate) fn try_reward_mixnode(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    mix_identity: IdentityKey,
-    uptime: u32,
-    rewarding_interval_nonce: u32,
-) -> Result<Response, ContractError> {
-    let state = config_read(deps.storage).load()?;
-
-    // check if this is executed by the permitted validator, if not reject the transaction
-    if info.sender != state.rewarding_validator_address {
-        return Err(ContractError::Unauthorized);
-    }
-
-    if !state.rewarding_in_progress {
-        return Err(ContractError::RewardingNotInProgress);
-    }
-
-    // make sure the transaction is sent for the correct rewarding interval
-    if rewarding_interval_nonce != state.latest_rewarding_interval_nonce {
-        return Err(ContractError::InvalidRewardingIntervalNonce {
-            received: rewarding_interval_nonce,
-            expected: state.latest_rewarding_interval_nonce,
-        });
-    }
-
-    // check if the mixnode hasn't been rewarded in this rewarding interval already
-    if rewarded_mixnodes_read(deps.storage, rewarding_interval_nonce)
-        .may_load(mix_identity.as_bytes())?
-        .is_some()
-    {
-        return Err(ContractError::MixnodeAlreadyRewarded {
-            identity: mix_identity,
-        });
-    }
-
-    // optimisation for uptime being 0. No rewards will be given so just terminate here
-    if uptime == 0 {
-        rewarded_mixnodes(deps.storage, rewarding_interval_nonce)
-            .save(mix_identity.as_bytes(), &Default::default())?;
-        return Ok(Response {
-            submessages: vec![],
-            messages: vec![],
-            attributes: vec![
-                attr("bond increase", Uint128(0)),
-                attr("total delegation increase", Uint128(0)),
-            ],
-            data: None,
-        });
-    }
-
-    // check if the bond even exists
-    let mut current_bond = match mixnodes_read(deps.storage).load(mix_identity.as_bytes()) {
-        Ok(bond) => bond,
-        Err(_) => {
-            return Ok(Response {
-                attributes: vec![attr("result", "bond not found")],
-                ..Default::default()
-            });
-        }
-    };
-
-    let mut node_reward = Uint128(0);
-    let mut total_delegation_reward = Uint128(0);
-
-    // update current bond with the reward given to the node and the delegators
-    // if it has been bonded for long enough
-    if current_bond.block_height + MINIMUM_BLOCK_AGE_FOR_REWARDING <= env.block.height {
-        let bond_reward_rate = state.mixnode_epoch_bond_reward;
-        let delegation_reward_rate = state.mixnode_epoch_delegation_reward;
-        let bond_scaled_reward_rate = scale_reward_by_uptime(bond_reward_rate, uptime)?;
-        let delegation_scaled_reward_rate = scale_reward_by_uptime(delegation_reward_rate, uptime)?;
-
-        total_delegation_reward = increase_mix_delegated_stakes(
-            deps.storage,
-            &mix_identity,
-            delegation_scaled_reward_rate,
-            env.block.height,
-        )?;
-
-        node_reward = current_bond.bond_amount.amount * bond_scaled_reward_rate;
-        current_bond.bond_amount.amount += node_reward;
-        current_bond.total_delegation.amount += total_delegation_reward;
-        mixnodes(deps.storage).save(mix_identity.as_bytes(), &current_bond)?;
-    }
-
-    rewarded_mixnodes(deps.storage, rewarding_interval_nonce)
-        .save(mix_identity.as_bytes(), &Default::default())?;
-
-    Ok(Response {
-        submessages: vec![],
-        messages: vec![],
-        attributes: vec![
-            attr("bond increase", node_reward),
-            attr("total delegation increase", total_delegation_reward),
-        ],
-        data: None,
-    })
-}
-
-pub(crate) fn try_reward_mixnode_v2(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    mix_identity: IdentityKey,
-    params: NodeRewardParams,
-    rewarding_interval_nonce: u32,
-) -> Result<Response, ContractError> {
-    let state = config_read(deps.storage).load()?;
-
-    // check if this is executed by the permitted validator, if not reject the transaction
-    if info.sender != state.rewarding_validator_address {
-        return Err(ContractError::Unauthorized);
-    }
-
-    if !state.rewarding_in_progress {
-        return Err(ContractError::RewardingNotInProgress);
-    }
-
-    // make sure the transaction is sent for the correct rewarding interval
-    if rewarding_interval_nonce != state.latest_rewarding_interval_nonce {
-        return Err(ContractError::InvalidRewardingIntervalNonce {
-            received: rewarding_interval_nonce,
-            expected: state.latest_rewarding_interval_nonce,
-        });
-    }
-
-    // check if the mixnode hasn't been rewarded in this rewarding interval already
-    if rewarded_mixnodes_read(deps.storage, rewarding_interval_nonce)
-        .may_load(mix_identity.as_bytes())?
-        .is_some()
-    {
-        return Err(ContractError::MixnodeAlreadyRewarded {
-            identity: mix_identity,
-        });
-    }
-
-    // check if the bond even exists
-    let mut current_bond = match mixnodes_read(deps.storage).load(mix_identity.as_bytes()) {
-        Ok(bond) => bond,
-        Err(_) => {
-            return Ok(Response {
-                attributes: vec![attr("result", "bond not found")],
-                ..Default::default()
-            });
-        }
-    };
-
-    let mut reward_params = params;
-
-    reward_params.set_reward_blockstamp(env.block.height);
-
-    let reward_result = current_bond.reward(&reward_params);
-
-    // Omitting the price per packet function now, it follows that base operator reward is the node_reward
-    let operator_reward = current_bond.operator_reward(&reward_params);
-
-    let total_delegation_reward =
-        increase_mix_delegated_stakes_v2(deps.storage, &current_bond, &reward_params)?;
-
-    // update current bond with the reward given to the node and the delegators
-    // if it has been bonded for long enough
-    if current_bond.block_height + MINIMUM_BLOCK_AGE_FOR_REWARDING
-        <= reward_params.reward_blockstamp()
-    {
-        current_bond.bond_amount.amount += Uint128(operator_reward);
-        current_bond.total_delegation.amount += total_delegation_reward;
-        mixnodes(deps.storage).save(mix_identity.as_bytes(), &current_bond)?;
-        decr_reward_pool(Uint128(operator_reward), deps.storage)?;
-        decr_reward_pool(total_delegation_reward, deps.storage)?;
-    }
-
-    rewarded_mixnodes(deps.storage, rewarding_interval_nonce)
-        .save(mix_identity.as_bytes(), &Default::default())?;
-
-    Ok(Response {
-        submessages: vec![],
-        messages: vec![],
-        attributes: vec![
-            attr("bond increase", reward_result.reward()),
-            attr("total delegation increase", total_delegation_reward),
-        ],
-        data: None,
-    })
-}
-pub(crate) fn try_finish_mixnode_rewarding(
-    deps: DepsMut,
-    info: MessageInfo,
-    rewarding_interval_nonce: u32,
-) -> Result<Response, ContractError> {
-    let mut state = config_read(deps.storage).load()?;
-
-    // check if this is executed by the permitted validator, if not reject the transaction
-    if info.sender != state.rewarding_validator_address {
-        return Err(ContractError::Unauthorized);
-    }
-
-    if !state.rewarding_in_progress {
-        return Err(ContractError::RewardingNotInProgress);
-    }
-
-    // make sure the validator is in sync with the contract state
-    if rewarding_interval_nonce != state.latest_rewarding_interval_nonce {
-        return Err(ContractError::InvalidRewardingIntervalNonce {
-            received: rewarding_interval_nonce,
-            expected: state.latest_rewarding_interval_nonce,
-        });
-    }
-
-    state.rewarding_in_progress = false;
-    config(deps.storage).save(&state)?;
-
-    Ok(Response::new())
-}
-
-fn validate_delegation_stake(delegation: &[Coin]) -> Result<(), ContractError> {
-    // check if anything was put as delegation
-    if delegation.is_empty() {
-        return Err(ContractError::EmptyDelegation);
-    }
-
-    if delegation.len() > 1 {
-        return Err(ContractError::MultipleDenoms);
-    }
-
-    // check that the denomination is correct
-    if delegation[0].denom != DENOM {
-        return Err(ContractError::WrongDenom {});
-    }
-
-    // check that we have provided a non-zero amount in the delegation
-    if delegation[0].amount.is_zero() {
-        return Err(ContractError::EmptyDelegation);
-    }
-
-    Ok(())
-}
-
-pub(crate) fn try_delegate_to_mixnode(
-    deps: DepsMut,
-    env: Env,
-    info: MessageInfo,
-    mix_identity: IdentityKey,
-) -> Result<Response, ContractError> {
-    // check if the delegation contains any funds of the appropriate denomination
-    validate_delegation_stake(&info.funds)?;
-
-    // check if the target node actually exists
-    let mut current_bond = match mixnodes_read(deps.storage).load(mix_identity.as_bytes()) {
-        Ok(bond) => bond,
-        Err(_) => {
-            return Err(ContractError::MixNodeBondNotFound {
-                identity: mix_identity,
-            });
-        }
-    };
-
-    let amount = info.funds[0].amount;
-
-    // update total_delegation of this node
-    current_bond.total_delegation.amount += info.funds[0].amount;
-    mixnodes(deps.storage).save(mix_identity.as_bytes(), &current_bond)?;
-
-    let mut delegation_bucket = mix_delegations(deps.storage, &mix_identity);
-    let sender_bytes = info.sender.as_bytes();
-
-    // write the delegation
-    let new_amount = match delegation_bucket.may_load(sender_bytes)? {
-        Some(existing_delegation) => existing_delegation.amount + amount,
-        None => amount,
-    };
-    // the block height is reset, if it existed
-    let new_delegation = RawDelegationData::new(new_amount, env.block.height);
-    delegation_bucket.save(sender_bytes, &new_delegation)?;
-
-    reverse_mix_delegations(deps.storage, &info.sender).save(mix_identity.as_bytes(), &())?;
-
-    Ok(Response::default())
-}
-
-pub(crate) fn try_remove_delegation_from_mixnode(
-    deps: DepsMut,
-    info: MessageInfo,
-    mix_identity: IdentityKey,
-) -> Result<Response, ContractError> {
-    let mut delegation_bucket = mix_delegations(deps.storage, &mix_identity);
-    let sender_bytes = info.sender.as_bytes();
-    match delegation_bucket.may_load(sender_bytes)? {
-        Some(delegation) => {
-            // remove delegation from the buckets
-            delegation_bucket.remove(sender_bytes);
-            reverse_mix_delegations(deps.storage, &info.sender).remove(mix_identity.as_bytes());
-
-            // send delegated funds back to the delegation owner
-            let messages = vec![BankMsg::Send {
-                to_address: info.sender.to_string(),
-                amount: coins(delegation.amount.u128(), DENOM),
-            }
-            .into()];
-
-            // update total_delegation of this node
-            let mut mixnodes_bucket = mixnodes(deps.storage);
-            // in some rare cases the mixnode bond might no longer exist as the node unbonded
-            // before delegation was removed. that is fine
-            if let Some(mut existing_bond) = mixnodes_bucket.may_load(mix_identity.as_bytes())? {
-                // we should NEVER underflow here, if we do, it means we have some serious error in our logic
-                existing_bond.total_delegation.amount = existing_bond
-                    .total_delegation
-                    .amount
-                    .checked_sub(delegation.amount)
-                    .unwrap();
-                mixnodes_bucket.save(mix_identity.as_bytes(), &existing_bond)?;
-            }
-
-            Ok(Response {
-                submessages: Vec::new(),
-                messages,
-                attributes: Vec::new(),
-                data: None,
-            })
-        }
-        None => Err(ContractError::NoMixnodeDelegationFound {
-            identity: mix_identity,
-            address: info.sender,
-        }),
-    }
-}
-
 #[cfg(test)]
 pub mod tests {
     use super::*;
+    use crate::bonding_gateways::transactions::try_add_gateway;
+    use crate::bonding_gateways::transactions::validate_gateway_bond;
+    use crate::bonding_mixnodes::transactions::try_add_mixnode;
+    use crate::bonding_mixnodes::transactions::validate_mixnode_bond;
     use crate::contract::{
         execute, query, DEFAULT_SYBIL_RESISTANCE_PERCENT, INITIAL_DEFAULT_EPOCH_LENGTH,
         INITIAL_GATEWAY_BOND, INITIAL_MIXNODE_BOND, INITIAL_MIXNODE_BOND_REWARD_RATE,
         INITIAL_MIXNODE_DELEGATION_REWARD_RATE,
     };
+    use crate::delegating_mixnodes::transactions::try_delegate_to_mixnode;
     use crate::helpers::calculate_epoch_reward_rate;
+    use crate::helpers::scale_reward_by_uptime;
+    use crate::helpers::Delegations;
+    use crate::mixnet_params::transactions::try_update_state_params;
+    use crate::queries::tests::store_n_mix_delegations;
     use crate::queries::DELEGATION_PAGE_DEFAULT_LIMIT;
+    use crate::rewards::transactions::{
+        try_begin_mixnode_rewarding, try_finish_mixnode_rewarding, try_reward_mixnode,
+        try_reward_mixnode_v2,
+    };
     use crate::storage::{layer_distribution_read, mix_delegations_read, read_mixnode_bond};
     use crate::support::tests::helpers;
     use crate::support::tests::helpers::{
         add_mixnode, good_gateway_bond, good_mixnode_bond, mix_node_fixture, raw_delegation_fixture,
     };
+    use cosmwasm_std::attr;
     use cosmwasm_std::testing::{mock_env, mock_info};
+    use cosmwasm_std::Decimal;
     use cosmwasm_std::{coin, coins, from_binary, Addr, Uint128};
+    use mixnet_contract::mixnode::NodeRewardParams;
+    use mixnet_contract::Gateway;
+    use mixnet_contract::MixNode;
+    use mixnet_contract::MixNodeBond;
+    use mixnet_contract::StateParams;
     use mixnet_contract::{
         ExecuteMsg, LayerDistribution, PagedGatewayResponse, PagedMixnodeResponse, QueryMsg,
         UnpackedDelegation,
     };
-    use queries::tests::store_n_mix_delegations;
 
     #[test]
     fn validating_mixnode_bond() {
@@ -1692,6 +989,7 @@ pub mod tests {
     #[cfg(test)]
     mod beginning_mixnode_rewarding {
         use super::*;
+        use crate::rewards::transactions::try_begin_mixnode_rewarding;
 
         #[test]
         fn can_only_be_called_by_specified_validator_address() {
@@ -1861,6 +1159,9 @@ pub mod tests {
     #[cfg(test)]
     mod finishing_mixnode_rewarding {
         use super::*;
+        use crate::rewards::transactions::{
+            try_begin_mixnode_rewarding, try_finish_mixnode_rewarding,
+        };
 
         #[test]
         fn can_only_be_called_by_specified_validator_address() {
@@ -2469,953 +1770,6 @@ pub mod tests {
         );
     }
 
-    #[cfg(test)]
-    mod delegation_stake_validation {
-        use super::*;
-        use cosmwasm_std::coin;
-
-        #[test]
-        fn stake_cant_be_empty() {
-            assert_eq!(
-                Err(ContractError::EmptyDelegation),
-                validate_delegation_stake(&[])
-            )
-        }
-
-        #[test]
-        fn stake_must_have_single_coin_type() {
-            assert_eq!(
-                Err(ContractError::MultipleDenoms),
-                validate_delegation_stake(&[coin(123, DENOM), coin(123, "BTC"), coin(123, "DOGE")])
-            )
-        }
-
-        #[test]
-        fn stake_coin_must_be_of_correct_type() {
-            assert_eq!(
-                Err(ContractError::WrongDenom {}),
-                validate_delegation_stake(&[coin(123, "DOGE")])
-            )
-        }
-
-        #[test]
-        fn stake_coin_must_have_value_greater_than_zero() {
-            assert_eq!(
-                Err(ContractError::EmptyDelegation),
-                validate_delegation_stake(&[coin(0, DENOM)])
-            )
-        }
-
-        #[test]
-        fn stake_can_have_any_positive_value() {
-            // this might change in the future, but right now an arbitrary (positive) value can be delegated
-            assert!(validate_delegation_stake(&[coin(1, DENOM)]).is_ok());
-            assert!(validate_delegation_stake(&[coin(123, DENOM)]).is_ok());
-            assert!(validate_delegation_stake(&[coin(10000000000, DENOM)]).is_ok());
-        }
-    }
-
-    #[cfg(test)]
-    mod mix_stake_delegation {
-        use super::*;
-        use crate::storage::mix_delegations_read;
-        use crate::support::tests::helpers::add_mixnode;
-
-        #[test]
-        fn fails_if_node_doesnt_exist() {
-            let mut deps = helpers::init_contract();
-            assert_eq!(
-                Err(ContractError::MixNodeBondNotFound {
-                    identity: "non-existent-mix-identity".into()
-                }),
-                try_delegate_to_mixnode(
-                    deps.as_mut(),
-                    mock_env(),
-                    mock_info("sender", &coins(123, DENOM)),
-                    "non-existent-mix-identity".into()
-                )
-            );
-        }
-
-        #[test]
-        fn succeeds_for_existing_node() {
-            let mut deps = helpers::init_contract();
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            let delegation = coin(123, DENOM);
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &vec![delegation.clone()]),
-                identity.clone()
-            )
-            .is_ok());
-
-            assert_eq!(
-                RawDelegationData::new(delegation.amount, mock_env().block.height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .load(identity.as_bytes())
-                    .is_ok()
-            );
-
-            // node's "total_delegation" is increased
-            assert_eq!(
-                delegation,
-                mixnodes_read(&deps.storage)
-                    .load(identity.as_bytes())
-                    .unwrap()
-                    .total_delegation
-            )
-        }
-
-        #[test]
-        fn fails_if_node_unbonded() {
-            let mut deps = helpers::init_contract();
-
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            try_remove_mixnode(deps.as_mut(), mock_info(mixnode_owner, &[])).unwrap();
-
-            assert_eq!(
-                Err(ContractError::MixNodeBondNotFound {
-                    identity: identity.clone()
-                }),
-                try_delegate_to_mixnode(
-                    deps.as_mut(),
-                    mock_env(),
-                    mock_info(delegation_owner.as_str(), &coins(123, DENOM)),
-                    identity
-                )
-            );
-        }
-
-        #[test]
-        fn succeeds_if_node_rebonded() {
-            let mut deps = helpers::init_contract();
-
-            let mixnode_owner = "bob";
-            add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            try_remove_mixnode(deps.as_mut(), mock_info(mixnode_owner, &[])).unwrap();
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation = coin(123, DENOM);
-            let delegation_owner = Addr::unchecked("sender");
-
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &vec![delegation.clone()]),
-                identity.clone()
-            )
-            .is_ok());
-
-            assert_eq!(
-                RawDelegationData::new(delegation.amount, mock_env().block.height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .load(identity.as_bytes())
-                    .is_ok()
-            );
-
-            // node's "total_delegation" is increased
-            assert_eq!(
-                delegation,
-                mixnodes_read(&deps.storage)
-                    .load(identity.as_bytes())
-                    .unwrap()
-                    .total_delegation
-            )
-        }
-
-        #[test]
-        fn is_possible_for_an_already_delegated_node() {
-            let mut deps = helpers::init_contract();
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            let delegation1 = coin(100, DENOM);
-            let delegation2 = coin(50, DENOM);
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &vec![delegation1.clone()]),
-                identity.clone(),
-            )
-            .unwrap();
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &vec![delegation2.clone()]),
-                identity.clone(),
-            )
-            .unwrap();
-
-            assert_eq!(
-                RawDelegationData::new(
-                    delegation1.amount + delegation2.amount,
-                    mock_env().block.height
-                ),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .load(identity.as_bytes())
-                    .is_ok()
-            );
-
-            // node's "total_delegation" is sum of both
-            assert_eq!(
-                delegation1.amount + delegation2.amount,
-                mixnodes_read(&deps.storage)
-                    .load(identity.as_bytes())
-                    .unwrap()
-                    .total_delegation
-                    .amount
-            )
-        }
-
-        #[test]
-        fn block_height_is_updated_on_new_delegation() {
-            let mut deps = helpers::init_contract();
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-            let delegation = coin(100, DENOM);
-
-            let env1 = mock_env();
-            let mut env2 = mock_env();
-            let initial_height = env1.block.height;
-            let updated_height = initial_height + 42;
-            // second env has grown in block height
-            env2.block.height = updated_height;
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                env1,
-                mock_info(delegation_owner.as_str(), &vec![delegation.clone()]),
-                identity.clone(),
-            )
-            .unwrap();
-
-            assert_eq!(
-                RawDelegationData::new(delegation.amount, initial_height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                env2,
-                mock_info(delegation_owner.as_str(), &vec![delegation.clone()]),
-                identity.clone(),
-            )
-            .unwrap();
-
-            assert_eq!(
-                RawDelegationData::new(delegation.amount + delegation.amount, updated_height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-        }
-
-        #[test]
-        fn block_height_is_not_updated_on_different_delegator() {
-            let mut deps = helpers::init_contract();
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner1 = Addr::unchecked("sender1");
-            let delegation_owner2 = Addr::unchecked("sender2");
-            let delegation1 = coin(100, DENOM);
-            let delegation2 = coin(120, DENOM);
-
-            let env1 = mock_env();
-            let mut env2 = mock_env();
-            let initial_height = env1.block.height;
-            let second_height = initial_height + 42;
-            // second env has grown in block height
-            env2.block.height = second_height;
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                env1,
-                mock_info(delegation_owner1.as_str(), &vec![delegation1.clone()]),
-                identity.clone(),
-            )
-            .unwrap();
-
-            assert_eq!(
-                RawDelegationData::new(delegation1.amount, initial_height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner1.as_bytes())
-                    .unwrap()
-            );
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                env2,
-                mock_info(delegation_owner2.as_str(), &vec![delegation2.clone()]),
-                identity.clone(),
-            )
-            .unwrap();
-
-            assert_eq!(
-                RawDelegationData::new(delegation1.amount, initial_height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner1.as_bytes())
-                    .unwrap()
-            );
-            assert_eq!(
-                RawDelegationData::new(delegation2.amount, second_height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner2.as_bytes())
-                    .unwrap()
-            );
-        }
-
-        #[test]
-        fn is_disallowed_for_already_delegated_node_if_it_unbonded() {
-            let mut deps = helpers::init_contract();
-
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &coins(100, DENOM)),
-                identity.clone(),
-            )
-            .unwrap();
-
-            try_remove_mixnode(deps.as_mut(), mock_info(mixnode_owner, &[])).unwrap();
-
-            assert_eq!(
-                Err(ContractError::MixNodeBondNotFound {
-                    identity: identity.clone()
-                }),
-                try_delegate_to_mixnode(
-                    deps.as_mut(),
-                    mock_env(),
-                    mock_info(delegation_owner.as_str(), &coins(50, DENOM)),
-                    identity
-                )
-            );
-        }
-
-        #[test]
-        fn is_allowed_for_multiple_nodes() {
-            let mut deps = helpers::init_contract();
-            let mixnode_owner1 = "bob";
-            let mixnode_owner2 = "fred";
-            let identity1 = add_mixnode(mixnode_owner1, good_mixnode_bond(), &mut deps);
-            let identity2 = add_mixnode(mixnode_owner2, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &coins(123, DENOM)),
-                identity1.clone()
-            )
-            .is_ok());
-
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &coins(42, DENOM)),
-                identity2.clone()
-            )
-            .is_ok());
-
-            assert_eq!(
-                RawDelegationData::new(123u128.into(), mock_env().block.height),
-                mix_delegations_read(&deps.storage, &identity1)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .load(identity1.as_bytes())
-                    .is_ok()
-            );
-
-            assert_eq!(
-                RawDelegationData::new(42u128.into(), mock_env().block.height),
-                mix_delegations_read(&deps.storage, &identity2)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .load(identity2.as_bytes())
-                    .is_ok()
-            );
-        }
-
-        #[test]
-        fn is_allowed_by_multiple_users() {
-            let mut deps = helpers::init_contract();
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-
-            let delegation1 = coin(123, DENOM);
-            let delegation2 = coin(234, DENOM);
-
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info("sender1", &vec![delegation1.clone()]),
-                identity.clone()
-            )
-            .is_ok());
-
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info("sender2", &vec![delegation2.clone()]),
-                identity.clone()
-            )
-            .is_ok());
-
-            // node's "total_delegation" is sum of both
-            assert_eq!(
-                delegation1.amount + delegation2.amount,
-                mixnodes_read(&deps.storage)
-                    .load(identity.as_bytes())
-                    .unwrap()
-                    .total_delegation
-                    .amount
-            )
-        }
-
-        #[test]
-        fn delegation_is_not_removed_if_node_unbonded() {
-            let mut deps = helpers::init_contract();
-
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &coins(100, DENOM)),
-                identity.clone(),
-            )
-            .unwrap();
-
-            try_remove_mixnode(deps.as_mut(), mock_info(mixnode_owner, &[])).unwrap();
-
-            assert_eq!(
-                RawDelegationData::new(100u128.into(), mock_env().block.height),
-                mix_delegations_read(&deps.storage, &identity)
-                    .load(delegation_owner.as_bytes())
-                    .unwrap()
-            );
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .load(identity.as_bytes())
-                    .is_ok()
-            );
-        }
-    }
-
-    #[cfg(test)]
-    mod removing_mix_stake_delegation {
-        use super::*;
-        use crate::storage::mix_delegations_read;
-        use crate::support::tests::helpers::add_mixnode;
-
-        #[test]
-        fn fails_if_delegation_never_existed() {
-            let mut deps = helpers::init_contract();
-
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            assert_eq!(
-                Err(ContractError::NoMixnodeDelegationFound {
-                    identity: identity.clone(),
-                    address: delegation_owner.clone(),
-                }),
-                try_remove_delegation_from_mixnode(
-                    deps.as_mut(),
-                    mock_info(delegation_owner.as_str(), &[]),
-                    identity,
-                )
-            );
-        }
-
-        #[test]
-        fn succeeds_if_delegation_existed() {
-            let mut deps = helpers::init_contract();
-
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &coins(100, DENOM)),
-                identity.clone(),
-            )
-            .unwrap();
-
-            assert_eq!(
-                Ok(Response {
-                    submessages: vec![],
-                    messages: vec![BankMsg::Send {
-                        to_address: delegation_owner.clone().into(),
-                        amount: coins(100, DENOM),
-                    }
-                    .into()],
-                    attributes: Vec::new(),
-                    data: None,
-                }),
-                try_remove_delegation_from_mixnode(
-                    deps.as_mut(),
-                    mock_info(delegation_owner.as_str(), &[]),
-                    identity.clone(),
-                )
-            );
-
-            assert!(mix_delegations_read(&deps.storage, &identity)
-                .may_load(delegation_owner.as_bytes())
-                .unwrap()
-                .is_none());
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .may_load(identity.as_bytes())
-                    .unwrap()
-                    .is_none()
-            );
-
-            // and total delegation is cleared
-            assert_eq!(
-                Uint128::zero(),
-                mixnodes_read(&deps.storage)
-                    .load(identity.as_bytes())
-                    .unwrap()
-                    .total_delegation
-                    .amount
-            )
-        }
-
-        #[test]
-        fn succeeds_if_delegation_existed_even_if_node_unbonded() {
-            let mut deps = helpers::init_contract();
-
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner = Addr::unchecked("sender");
-
-            try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner.as_str(), &coins(100, DENOM)),
-                identity.clone(),
-            )
-            .unwrap();
-
-            try_remove_mixnode(deps.as_mut(), mock_info(mixnode_owner, &[])).unwrap();
-
-            assert_eq!(
-                Ok(Response {
-                    submessages: vec![],
-                    messages: vec![BankMsg::Send {
-                        to_address: delegation_owner.clone().into(),
-                        amount: coins(100, DENOM),
-                    }
-                    .into()],
-                    attributes: Vec::new(),
-                    data: None,
-                }),
-                try_remove_delegation_from_mixnode(
-                    deps.as_mut(),
-                    mock_info(delegation_owner.as_str(), &[]),
-                    identity.clone(),
-                )
-            );
-
-            assert!(mix_delegations_read(&deps.storage, &identity)
-                .may_load(delegation_owner.as_bytes())
-                .unwrap()
-                .is_none());
-            assert!(
-                reverse_mix_delegations_read(&deps.storage, &delegation_owner)
-                    .may_load(identity.as_bytes())
-                    .unwrap()
-                    .is_none()
-            );
-        }
-
-        #[test]
-        fn total_delegation_is_preserved_if_only_some_undelegate() {
-            let mut deps = helpers::init_contract();
-            let mixnode_owner = "bob";
-            let identity = add_mixnode(mixnode_owner, good_mixnode_bond(), &mut deps);
-            let delegation_owner1 = Addr::unchecked("sender1");
-            let delegation_owner2 = Addr::unchecked("sender2");
-
-            let delegation1 = coin(123, DENOM);
-            let delegation2 = coin(234, DENOM);
-
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner1.as_str(), &vec![delegation1.clone()]),
-                identity.clone()
-            )
-            .is_ok());
-
-            assert!(try_delegate_to_mixnode(
-                deps.as_mut(),
-                mock_env(),
-                mock_info(delegation_owner2.as_str(), &vec![delegation2.clone()]),
-                identity.clone()
-            )
-            .is_ok());
-
-            // sender1 undelegates
-            try_remove_delegation_from_mixnode(
-                deps.as_mut(),
-                mock_info(delegation_owner1.as_str(), &[]),
-                identity.clone(),
-            )
-            .unwrap();
-
-            // but total delegation should still equal to what sender2 sent
-            // node's "total_delegation" is sum of both
-            assert_eq!(
-                delegation2,
-                mixnodes_read(&deps.storage)
-                    .load(identity.as_bytes())
-                    .unwrap()
-                    .total_delegation
-            )
-        }
-    }
-
-    #[test]
-    fn delegators_on_mix_node_reward_rate() {
-        let mut deps = helpers::init_contract();
-        let mut env = mock_env();
-        let current_state = config_read(deps.as_mut().storage).load().unwrap();
-        let rewarding_validator_address = current_state.rewarding_validator_address;
-
-        let initial_mix_bond = 100_000000;
-        let initial_delegation1 = 50000; // will see single digits rewards
-        let initial_delegation2 = 100; // won't see any rewards due to such a small delegation
-        let initial_delegation3 = 100000_000000; // will see big proper rewards
-
-        let node_owner = "node-owner";
-        let identity = add_mixnode(node_owner, good_mixnode_bond(), &mut deps);
-
-        mix_delegations(&mut deps.storage, &identity)
-            .save(
-                b"delegator1",
-                &RawDelegationData::new(initial_delegation1.into(), env.block.height),
-            )
-            .unwrap();
-        mix_delegations(&mut deps.storage, &identity)
-            .save(
-                b"delegator2",
-                &RawDelegationData::new(initial_delegation2.into(), env.block.height),
-            )
-            .unwrap();
-        mix_delegations(&mut deps.storage, &identity)
-            .save(
-                b"delegator3",
-                &RawDelegationData::new(initial_delegation3.into(), env.block.height),
-            )
-            .unwrap();
-
-        env.block.height += 2 * MINIMUM_BLOCK_AGE_FOR_REWARDING;
-
-        let bond_reward = current_state.mixnode_epoch_bond_reward;
-        let delegation_reward = current_state.mixnode_epoch_delegation_reward;
-
-        // the node's bond and delegations are correctly increased and scaled by uptime
-        // if node was 100% up, it will get full epoch reward
-        let expected_mix_reward = Uint128(initial_mix_bond) * bond_reward;
-        let expected_delegation1_reward = Uint128(initial_delegation1) * delegation_reward;
-        let expected_delegation2_reward = Uint128(initial_delegation2) * delegation_reward;
-        let expected_delegation3_reward = Uint128(initial_delegation3) * delegation_reward;
-
-        let expected_bond = expected_mix_reward + Uint128(initial_mix_bond);
-        let expected_delegation1 = expected_delegation1_reward + Uint128(initial_delegation1);
-        let expected_delegation2 = expected_delegation2_reward + Uint128(initial_delegation2);
-        let expected_delegation3 = expected_delegation3_reward + Uint128(initial_delegation3);
-
-        let info = mock_info(rewarding_validator_address.as_ref(), &[]);
-        try_begin_mixnode_rewarding(deps.as_mut(), env.clone(), info.clone(), 1).unwrap();
-        let res = try_reward_mixnode(
-            deps.as_mut(),
-            env.clone(),
-            info.clone(),
-            identity.clone(),
-            100,
-            1,
-        )
-        .unwrap();
-        try_finish_mixnode_rewarding(deps.as_mut(), info, 1).unwrap();
-
-        assert_eq!(
-            expected_bond,
-            read_mixnode_bond(deps.as_ref().storage, identity.as_bytes()).unwrap()
-        );
-
-        assert_eq!(
-            expected_delegation1,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator1".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            expected_delegation2,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator2".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            expected_delegation3,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator3".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            vec![
-                attr("bond increase", expected_mix_reward),
-                attr(
-                    "total delegation increase",
-                    expected_delegation1_reward
-                        + expected_delegation2_reward
-                        + expected_delegation3_reward
-                ),
-            ],
-            res.attributes
-        );
-
-        // if node was 20% up, it will get 1/5th of epoch reward
-        let scaled_bond_reward = scale_reward_by_uptime(bond_reward, 20).unwrap();
-        let scaled_delegation_reward = scale_reward_by_uptime(delegation_reward, 20).unwrap();
-
-        let expected_mix_reward = expected_bond * scaled_bond_reward;
-        let expected_delegation1_reward = expected_delegation1 * scaled_delegation_reward;
-        let expected_delegation2_reward = expected_delegation2 * scaled_delegation_reward;
-        let expected_delegation3_reward = expected_delegation3 * scaled_delegation_reward;
-
-        let expected_bond = expected_mix_reward + expected_bond;
-        let expected_delegation1 = expected_delegation1_reward + expected_delegation1;
-        let expected_delegation2 = expected_delegation2_reward + expected_delegation2;
-        let expected_delegation3 = expected_delegation3_reward + expected_delegation3;
-
-        let info = mock_info(rewarding_validator_address.as_ref(), &[]);
-        try_begin_mixnode_rewarding(deps.as_mut(), env.clone(), info.clone(), 2).unwrap();
-        let res = try_reward_mixnode(
-            deps.as_mut(),
-            env.clone(),
-            info.clone(),
-            identity.clone(),
-            20,
-            2,
-        )
-        .unwrap();
-        try_finish_mixnode_rewarding(deps.as_mut(), info, 2).unwrap();
-
-        assert_eq!(
-            expected_bond,
-            read_mixnode_bond(deps.as_ref().storage, identity.as_bytes()).unwrap()
-        );
-
-        assert_eq!(
-            expected_delegation1,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator1".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            expected_delegation2,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator2".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            expected_delegation3,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator3".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            vec![
-                attr("bond increase", expected_mix_reward),
-                attr(
-                    "total delegation increase",
-                    expected_delegation1_reward
-                        + expected_delegation2_reward
-                        + expected_delegation3_reward
-                ),
-            ],
-            res.attributes
-        );
-
-        // if the node was 0% up, nobody will get any rewards
-        let info = mock_info(rewarding_validator_address.as_ref(), &[]);
-        try_begin_mixnode_rewarding(deps.as_mut(), env.clone(), info.clone(), 3).unwrap();
-        let res = try_reward_mixnode(
-            deps.as_mut(),
-            env.clone(),
-            info.clone(),
-            identity.clone(),
-            0,
-            3,
-        )
-        .unwrap();
-        try_finish_mixnode_rewarding(deps.as_mut(), info, 3).unwrap();
-
-        assert_eq!(
-            expected_bond,
-            read_mixnode_bond(deps.as_ref().storage, identity.as_bytes()).unwrap()
-        );
-
-        assert_eq!(
-            expected_delegation1,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator1".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            expected_delegation2,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator2".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            expected_delegation3,
-            mix_delegations_read(deps.as_ref().storage, &identity)
-                .load("delegator3".as_bytes())
-                .unwrap()
-                .amount
-        );
-
-        assert_eq!(
-            vec![
-                attr("bond increase", Uint128(0)),
-                attr("total delegation increase", Uint128(0)),
-            ],
-            res.attributes
-        );
-    }
-
-    #[test]
-    fn multiple_page_delegations() {
-        let mut deps = helpers::init_contract();
-        let node_identity: IdentityKey = "foo".into();
-
-        store_n_mix_delegations(
-            DELEGATION_PAGE_DEFAULT_LIMIT * 10,
-            &mut deps.storage,
-            &node_identity,
-        );
-        let mix_bucket = all_mix_delegations_read::<RawDelegationData>(&deps.storage);
-        let mix_delegations =
-            Delegations::new(mix_bucket).collect::<Vec<UnpackedDelegation<RawDelegationData>>>();
-        assert_eq!(
-            DELEGATION_PAGE_DEFAULT_LIMIT * 10,
-            mix_delegations.len() as u32
-        );
-    }
-
-    #[cfg(test)]
-    mod finding_old_delegations {
-        use super::*;
-
-        #[test]
-        fn when_there_werent_any() {
-            let deps = helpers::init_contract();
-
-            let node_identity: IdentityKey = "nodeidentity".into();
-
-            let read_bucket = mix_delegations_read(&deps.storage, &node_identity);
-            let old_delegations = total_delegations(read_bucket).unwrap();
-
-            assert_eq!(Coin::new(0, DENOM), old_delegations);
-        }
-
-        #[test]
-        fn when_some_existed() {
-            let num_delegations = vec![
-                1,
-                5,
-                OLD_DELEGATIONS_CHUNK_SIZE - 1,
-                OLD_DELEGATIONS_CHUNK_SIZE,
-                OLD_DELEGATIONS_CHUNK_SIZE + 1,
-                OLD_DELEGATIONS_CHUNK_SIZE * 3,
-                OLD_DELEGATIONS_CHUNK_SIZE * 3 + 1,
-            ];
-
-            for delegations in num_delegations {
-                let mut deps = helpers::init_contract();
-
-                let node_identity: IdentityKey = "nodeidentity".into();
-
-                // delegate some stake
-                let mut write_bucket = mix_delegations(&mut deps.storage, &node_identity);
-                for i in 1..=delegations {
-                    let delegator = Addr::unchecked(format!("delegator{}", i));
-                    let delegation = raw_delegation_fixture(i as u128);
-                    write_bucket
-                        .save(delegator.as_bytes(), &delegation)
-                        .unwrap();
-                }
-
-                let read_bucket = mix_delegations_read(&deps.storage, &node_identity);
-                let old_delegations = total_delegations(read_bucket).unwrap();
-
-                let total_delegation = (1..=delegations as u128).into_iter().sum();
-                assert_eq!(Coin::new(total_delegation, DENOM), old_delegations);
-            }
-        }
-    }
-
     #[test]
     fn choose_layer_mix_node() {
         let mut deps = helpers::init_contract();
@@ -3437,7 +1791,7 @@ pub mod tests {
         assert_eq!(alice_node.mix_node.identity_key, "alice");
         assert_eq!(alice_node.layer, Layer::One);
         assert_eq!(bob_node.mix_node.identity_key, "bob");
-        assert_eq!(bob_node.layer, Layer::Two);
+        assert_eq!(bob_node.layer, mixnet_contract::Layer::Two);
     }
 
     #[test]
@@ -3466,7 +1820,7 @@ pub mod tests {
                 "alice",
                 &vec![Coin {
                     denom: DENOM.to_string(),
-                    amount: Uint128(10000_000_000),
+                    amount: Uint128(10_000_000_000),
                 }],
             ),
             MixNode {
@@ -3542,12 +1896,12 @@ pub mod tests {
         assert_eq!(mix1_delegator2_reward, U128::from_num(5638153));
 
         let pre_reward_bond = read_mixnode_bond(&deps.storage, b"alice").unwrap().u128();
-        assert_eq!(pre_reward_bond, 10000_000_000);
+        assert_eq!(pre_reward_bond, 10_000_000_000);
 
         let pre_reward_delegation = read_mixnode_delegation(&deps.storage, b"alice")
             .unwrap()
             .u128();
-        assert_eq!(pre_reward_delegation, 10000_000_000);
+        assert_eq!(pre_reward_delegation, 10_000_000_000);
 
         try_reward_mixnode_v2(deps.as_mut(), env, info, "alice".to_string(), params, 1).unwrap();
